fix view headers and create header buttons
