+++ src/assets/scss/components/header/_Header.scss
@@ -1,5 +1,7 @@
 .c-headers-options{
-
+  .p-2{
+    padding: 0px !important;
+  }
 }
 
 .c-edit-header {
