+++ src/pages/MyHeaders/MyHeadersPage.js
@@ -1,17 +1,25 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {
-  Row, Col, Alert,
+  Row, Col, Alert, Button,
   UncontrolledDropdown, DropdownItem, DropdownMenu, DropdownToggle,
 } from 'reactstrap';
 
+import { Link } from 'react-router-dom';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { history } from 'helpers/history';
+
 import CustomPagination from 'components/pagination/CustomPagination';
-import HeaderOptions from 'components/headers/HeaderOptions';
 import HeadersList from 'components/headers/HeadersList';
 
 import { ToastContainer } from 'react-toastify';
 import HomeUserPage from '../HomeUser/HomeUserPage';
 import 'react-toastify/dist/ReactToastify.css';
+
+const createNewHeader = (resetNewHeader) => {
+  history.replace('/new-header');
+  resetNewHeader();
+};
 
 const getOrderNameField = (text) => {
   switch (text) {
@@ -54,7 +62,18 @@ class MyHeadersPage extends React.Component {
     return (
       <HomeUserPage>
         <ToastContainer hideProgressBar position="bottom-right" />
-        <HeaderOptions resetNewHeader={resetNewHeader} />
+        <Row className="c-headers-options">
+          <Col sm="12" className="d-flex justify-content-end">
+            <div className="p-2">
+              <Link className="" to="/new-header">
+                <Button onClick={() => createNewHeader(resetNewHeader)}>
+                  <FontAwesomeIcon icon="plus" className="btn__icon" />
+                  Criar novo cabeçalho
+                </Button>
+              </Link>
+            </div>
+          </Col>
+        </Row>
 
         <div className="c-my-headers">
           <Row>
@@ -137,7 +156,6 @@ class MyHeadersPage extends React.Component {
       </HomeUserPage>);
   }
 }
-
 
 MyHeadersPage.propTypes = {
   match: PropTypes.shape({
